+++ test/IO.ts
@@ -1,6 +1,6 @@
 import * as IO from "fp-ts/IO"
-import { constant, constVoid } from "fp-ts/function"
-import { tap, once } from "../src/IO"
+import { constant, constFalse, constTrue, constVoid } from "fp-ts/function"
+import { tap, once, whenInvocationCount } from "../src/IO"
 import { add } from "../src/Number"
 import fc from "fast-check"
 
@@ -66,4 +66,55 @@ describe("IO", () => {
       expect(h(3)()).toBe(17)
     })
   })
+
+  /* eslint-disable */
+  describe("whenInvocationCount", () => {
+    const f = whenInvocationCount
+
+    it("always runs the effect on const true predicate", () => {
+      fc.assert(
+        fc.property(fc.integer({ min: 0, max: 99 }), x => {
+          let n = 0
+          const g: IO<void> = () => {
+            n++
+          }
+          const h = f(constTrue)(g)
+
+          for (let i = 0; i < x; i++) h()
+          return n === x
+        }),
+      )
+    })
+
+    it("never runs the effect on const false predicate", () => {
+      fc.assert(
+        fc.property(fc.integer({ min: 0, max: 99 }), x => {
+          let n = 0
+          const g: IO<void> = () => {
+            n++
+          }
+          const h = f(constFalse)(g)
+
+          for (let i = 0; i < x; i++) h()
+          return n === 0
+        }),
+      )
+    })
+
+    it("runs the effect whenever the predicate passes", () => {
+      fc.assert(
+        fc.property(fc.integer({ min: 0, max: 99 }), x => {
+          let n = 0
+          const g: IO<void> = () => {
+            n++
+          }
+          const h = f(n => n % 2 === 0)(g)
+
+          for (let i = 0; i < x; i++) h()
+          return n === Math.floor(x / 2)
+        }),
+      )
+    })
+  })
+  /* eslint-enable */
 })